move tests to __tests__ subdirs
